% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_ekr_ref.R
\docType{data}
\name{ekr_ref}
\alias{ekr_ref}
\title{EKR Reference Data}
\format{
A data frame with 3 columns:
\describe{
\item{watertype}{The water body type (character)}
\item{ekr}{The EKR value (numeric)}
\item{maatlat}{The part of the measurement value (character)}
}
}
\source{
deelmaatlat STOWA
}
\usage{
ekr_ref
}
\description{
EKR Reference Data
}
\keyword{datasets}
