% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_to_ekr.R
\name{to_ekr}
\alias{to_ekr}
\title{Convert measurement value to EKR}
\usage{
to_ekr(x, watertype, deelmaatlat)
}
\arguments{
\item{x}{The measurement value (chlfa in ug/l, waterdiepte_grens(met/zonder planten) in m, and bedekking in \%)}

\item{watertype}{The water body type}

\item{deelmaatlat}{The part of the measurement value}
}
\value{
The EKR value
}
\description{
This function converts a measurement value to EKR
}
\examples{
to_ekr(10.99, "M6b", "chlfa")
}
